+++ tests/Support/StubDateTimeImmutableType.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Vjik\CycleTypecast\Tests\DateTimeImmutable;
+namespace Vjik\CycleTypecast\Tests\Support;
 
 use DateTimeImmutable;
 use Vjik\CycleTypecast\DateTimeImmutable\DateTimeImmutableType;
 
 final class StubDateTimeImmutableType extends DateTimeImmutableType
 {
-    protected function toDatabaseValue(DateTimeImmutable $value)
+    protected function toDatabaseValue(DateTimeImmutable $value): mixed
     {
         return $value->getTimestamp();
     }