+++ tests/Support/StubUuidStringType.php
@@ -2,19 +2,19 @@
 
 declare(strict_types=1);
 
-namespace Vjik\CycleTypecast\Tests\UuidString;
+namespace Vjik\CycleTypecast\Tests\Support;
 
 use Ramsey\Uuid\UuidInterface;
 use Vjik\CycleTypecast\UuidString\UuidStringType;
 
 final class StubUuidStringType extends UuidStringType
 {
-    protected function toDatabaseValue(UuidInterface $value)
+    protected function toDatabaseValue(UuidInterface $value): mixed
     {
         return $value->toString();
     }
 
-    protected function toPhpValue($value): string
+    protected function toPhpValue(mixed $value): string
     {
         return $value;
     }