+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Cycle Typecast Change Log
 
+## 2.0.0 May 24, 2022
+
+- New #2: Add abstract typecast handler
+- Chg #2: Refactoring `Typecaster` (changed method names and returned values)
+- Chg #2: Raise the minimum PHP version to `^8.0` and Cycle ORM to `^2.1` 
+
 ## 1.0.0 December 1, 2021
 
 - Initial release.